ipo_ai/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
scikit-learn
requests
beautifulsoup4
apscheduler
joblib
selenium
webdriver-manager
pyyaml
